Need Fix Update totalPrice when Deleted

+++ BackEnd/controllers/cart.controller.js
@@ -53,26 +53,29 @@ const addToCart = async(req, res) => {
 
 const deleteFromCart = async(req , res) =>{
     try{
-        const {userId , cartItemId} = req.params;
+        const {userId , cartItemId} = req.params;      
 
-        let cart = await Cart.findOne({user : userId}).populate({
+        const cart = await Cart.findOne({user : userId}).populate({
             path: 'items',
             populate: {
                 path: 'book'
             }
         });
-
+        
+        const cartItem = await CartItem.findOne({_id: cartItemId});
+        if(!cartItem){
+            return res.status(404).json({message: 'Không có sản phẩm này trong giỏ hàng'});
+        }
+        await CartItem.findOneAndDelete({ _id: cartItemId });
         for (let i = 0; i < cart.items.length; i++) {
             const item = cart.items[i];
-            if (item._id.equals(cartItemId)) {
-                cart.totalPriceOriginal -= (item.book.priceOriginal*item.quantity);
-                cart.totalPriceFinal -= (item.book.priceFinal*item.quantity);
+            if (item && item._id.equals(cartItemId)) {
+                cart.totalPriceOriginal -= (item.book.priceOriginal * item.quantity);
+                cart.totalPriceFinal -= (item.book.priceFinal * item.quantity);
                 cart.items.splice(i, 1);
                 break;
             }
-        }
-        
-        await CartItem.findOneAndDelete({ _id: cartItemId });
+        }        
         await cart.save();
         
         res.status(200).json({ message: 'Xóa sản phẩm khỏi giỏ hàng thành công' });
@@ -132,7 +135,12 @@ const updateQuantityOfCartItem = async (req, res) => {
 const getCart = async(req, res) => {
     try {
         const userId = req.params.userId;
-        const cart = await Cart.findOne({ user: userId }).populate('items');
+        const cart = await Cart.findOne({ user: userId }).populate({
+            path: 'items',
+            populate: {
+                path: 'book'
+            }
+        });
 
         if (!cart) {
             return res.status(404).json({ message: 'Không tìm thấy giỏ hàng cho người dùng này' });
